chore: PaymentServiceTest 충돌 수정 및 테스트를 위한 정보 추가

+++ backend/src/main/java/com/pickgo/domain/payment/dto/PaymentConfirmRequest.java
@@ -6,5 +6,6 @@ import jakarta.validation.constraints.NotNull;
 public record PaymentConfirmRequest(
         @NotBlank String paymentKey,
         @NotBlank String orderId,
-        @NotNull Long amount
-) {}
+        @NotNull Integer amount
+) {
+}
